# backend/openapi.yaml
openapi: 3.1.0
info:
  title: RMG ERP System
  version: 1.0.0
  description: |
    Southern Apparels ERP System API
    Multi-Database Architecture with 5 separate PostgreSQL databases
    - Clients DB: Buyers, Suppliers, Contacts, Shipping, Banking
    - Samples DB: Styles, Variants, Samples, TNA, Plans, Operations, Materials, Colors, Sizes
    - Users DB: User management & Authentication
    - Orders DB: Order management
    - Merchandiser DB: Future merchandiser features
servers:
  - url: /
    description: Current server (auto-detected)
...
paths:
...
  /api/v1/users/:
    get:
      tags: [users]
      summary: List all users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              example:
                - id: 1
                  email: "admin@rmg.com"
                  username: "admin"
                  full_name: "System Admin"
                  department: "IT"
                  designation: "Administrator"
                  department_access: ["all"]
                  is_active: true
                  is_superuser: true
                  created_at: "2024-01-01T00:00:00Z"
    post:
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "merchandiser@rmg.com"
              username: "merchandiser1"
              full_name: "John Merchandiser"
              department: "Merchandising"
              designation: "Senior Merchandiser"
              department_access: ["samples", "orders", "buyers"]
              password: "SecurePass123"
      responses:
        '201':
          description: User created
  /api/v1/users/{user_id}:
    get:
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: User details
        '404': { description: Not found }
    put:
      tags: [users]
      summary: Update user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            example:
              full_name: "Senior Admin"
              designation: "IT Manager"
              is_active: true
      responses:
        '200':
          description: User updated
    delete:
      tags: [users]
      summary: Delete user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '204': { description: Deleted }

  # ============ OPERATIONS ============
...